+++ webpack.config.js
@@ -3,6 +3,7 @@ const path = require('path');
 module.exports = {
     entry: './src/streamerworm.js',
     mode: 'production',
+    devtool: 'source-map',
     output: {
         filename: 'main.js',
         path: path.resolve(__dirname, 'dist'),